GETDiscordBot/requirements.txt
discord.py==1.7.3
requests==2.26.0
quickchart.io==1.0.0
